requirements.txt
numpy==1.21.1
pandas==1.3.3
matplotlib
scikit-learn==0.24.2
nltk==3.5
streamlit==0.88.0
wordcloud
regex
oauthlib
pillow
streamlit
seaborn